% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnsariBradley.R
\name{AnsariBradley}
\alias{AnsariBradley}
\title{Ansari-Bradley Test}
\description{
Performs two sample Ansari-Bradley test on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::Two Sample Permutation Test} -> \code{LearnNonparam::Two Sample Wilcoxon Test} -> \code{Ansari-Bradley Test}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ansari-Bradley Test-new}{\code{AnsariBradley$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot_hist"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot_hist'><code>LearnNonparam::Permutation Test$plot_hist()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ansari-Bradley Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ansari-Bradley Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{AnsariBradley} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnsariBradley$new(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{parameters passed to \code{Wilcoxon$new()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{AnsariBradley} object.
}
}
}
