% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContingencyTableTest.R
\name{ContingencyTableTest}
\alias{ContingencyTableTest}
\alias{class.table}
\title{ContingencyTableTest Class}
\description{
Abstract class for tests on contingency tables.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{ContingencyTableTest}
}
