% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JonckheereTerpstra.R
\name{JonckheereTerpstra}
\alias{JonckheereTerpstra}
\title{Jonckheere-Terpstra Test}
\description{
Performs k sample Jonckheere-Terpstra Test on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::K Sample Permutation Test} -> \code{Jonckheere-Terpstra Test}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Jonckheere-Terpstra Test-new}{\code{JonckheereTerpstra$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot_hist"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot_hist'><code>LearnNonparam::Permutation Test$plot_hist()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Jonckheere-Terpstra Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Jonckheere-Terpstra Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{JonckheereTerpstra} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{JonckheereTerpstra$new(
  type = c("permu", "approx"),
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"permu"} (default) or \code{"approx"}.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{JonckheereTerpstra} object.
}
}
}
