% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KruskalWallis.R
\name{KruskalWallis}
\alias{KruskalWallis}
\title{Kruskal-Wallis Test}
\description{
Performs k sample Kruskal-Wallis Test on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::K Sample Permutation Test} -> \code{Kruskal-Wallis Test}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Kruskal-Wallis Test-new}{\code{KruskalWallis$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot'><code>LearnNonparam::Permutation Test$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="print"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-print'><code>LearnNonparam::Permutation Test$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Kruskal-Wallis Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Kruskal-Wallis Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{KruskalWallis} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KruskalWallis$new(
  type = c("permu", "approx"),
  n_permu = NULL,
  scoring = c("rank", "vw", "expon")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"permu"} (default) or \code{"approx"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}

\item{\code{scoring}}{a character string specifying which scoring system to be used, must be one of \code{"rank"} (default), \code{"vw"} or \code{"expon"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{KruskalWallis} object.
}
}
}
