% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSampleTest.R
\name{KSampleTest}
\alias{KSampleTest}
\title{KSampleTest Class}
\description{
This Test specializes \code{PermuTest} for k sample permutation tests. Note that it is not recommended to create objects of this class directly.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{KSampleTest}
}
