% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mean.R
\name{Mean}
\alias{Mean}
\alias{twosample.mean}
\title{Two Sample Test Based on Mean}
\description{
Performs mean based two sample permutation test on data vectors.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{Mean}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Mean-new}{\code{Mean$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="feed"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-feed'><code>LearnNonparam::PermuTest$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Mean-new"></a>}}
\if{latex}{\out{\hypertarget{method-Mean-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Mean} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Mean$new(alternative = c("two_sided", "less", "greater"), n_permu = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alternative}}{a character string specifying the alternative hypothesis.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Mean} object.
}
}
}
