% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompT.R
\name{MultiCompT}
\alias{MultiCompT}
\alias{multicomp.t}
\title{Multiple Comparison Based on t Statistic}
\description{
Performs t statistic based multiple comparison on data vectors.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:KSampleTest]{LearnNonparam::KSampleTest}} -> \code{\link[LearnNonparam:MultipleComparison]{LearnNonparam::MultipleComparison}} -> \code{MultiCompT}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MultiCompT-new}{\code{MultiCompT$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="test"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-test'><code>LearnNonparam::PermuTest$test()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MultiCompT-new"></a>}}
\if{latex}{\out{\hypertarget{method-MultiCompT-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MultiCompT} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiCompT$new(
  type = c("permu", "asymp"),
  method = c("bonferroni", "no_bonferroni"),
  conf_level = 0.95,
  n_permu = 0L,
  scoring = c("none", "rank", "vw", "expon")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate the p-value.}

\item{\code{conf_level}}{a numeric value between zero and one giving the family-wise confidence level to use.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If set to zero (default) then all permutations are used.}

\item{\code{scoring}}{a character string specifying which scoring system to be used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{MultiCompT} object.
}
}
}
