% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleComparison.R
\name{MultipleComparison}
\alias{MultipleComparison}
\title{MultipleComparison Class}
\description{
This class specializes \code{KSampleTest} for multiple comparisons. Note that it is not recommended to create objects of this class directly.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:KSampleTest]{LearnNonparam::KSampleTest}} -> \code{MultipleComparison}
}
