% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSampleTest.R
\name{OneSampleTest}
\alias{OneSampleTest}
\title{OneSampleTest Class}
\description{
Abstract class for one sample tests.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{OneSampleTest}
}
