% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Page.R
\name{Page}
\alias{Page}
\title{Page Test}
\description{
Performs Page test on data for a randomized complete block design.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::Randomized Complete Block Design} -> \code{Page Test}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Page Test-new}{\code{Page$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="ggplot"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-ggplot'><code>LearnNonparam::Permutation Test$ggplot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Page Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Page Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Page} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Page$new(
  type = c("permu", "approx"),
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"permu"} (default) or \code{"approx"}.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Page} object.
}
}
}
