% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairedDifference.R
\name{PairedDifference}
\alias{PairedDifference}
\alias{paired.difference}
\title{Paired Comparison Based on Differences}
\description{
Performs two sample signed score test on data vectors.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{\link[LearnNonparam:TwoSamplePairedTest]{LearnNonparam::TwoSamplePairedTest}} -> \code{PairedDifference}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PairedDifference-new}{\code{PairedDifference$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="test"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-test'><code>LearnNonparam::PermuTest$test()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PairedDifference-new"></a>}}
\if{latex}{\out{\hypertarget{method-PairedDifference-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PairedDifference} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PairedDifference$new(
  type = c("permu", "asymp"),
  method = c("with_zeros", "without_zeros"),
  correct = TRUE,
  alternative = c("two_sided", "less", "greater"),
  n_permu = 0L,
  scoring = c("none", "rank", "vw", "expon")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate the p-value.}

\item{\code{method}}{a character string specifying the method of ranking data in computing adjusted signed ranks for tied data, must be one of \code{"with_zeros"} (default) or \code{"ignore"}. Note that the data will be modified when this parameter is set to \code{"ignore"}.}

\item{\code{correct}}{a logical indicating whether to apply continuity correction in the normal approximation for the p-value when \code{scoring} is set to \code{"rank"}.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If set to zero (default) then all permutations are used.}

\item{\code{scoring}}{a character string specifying which scoring system to be used on the absolute differences.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{PairedDifference} object.
}
}
}
