% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt.R
\docType{data}
\name{pmts}
\alias{pmts}
\title{All tests available.}
\format{
An object of class \code{data.frame} with 22 rows and 2 columns.
}
\usage{
pmts
}
\description{
All tests available.
}
\keyword{datasets}
