% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RatioMeanDeviance.R
\name{RatioMeanDeviance}
\alias{RatioMeanDeviance}
\title{Ratio Mean Deviance Test}
\description{
Performs two sample ratio mean deviance test on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::Two Sample Permutation Test} -> \code{Ratio Mean Deviance Test}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ratio Mean Deviance Test-new}{\code{RatioMeanDeviance$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot'><code>LearnNonparam::Permutation Test$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="print"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-print'><code>LearnNonparam::Permutation Test$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ratio Mean Deviance Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ratio Mean Deviance Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{RatioMeanDeviance} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RatioMeanDeviance$new(
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{RatioMeanDeviance} object.
}
}
}
