% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCBD.R
\name{RCBD}
\alias{RCBD}
\title{RCBD Class}
\description{
Abstract class for randomized complete block design.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{RCBD}
}
