% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCBDANOVA.R
\name{RCBDANOVA}
\alias{RCBDANOVA}
\alias{rcbd.anova}
\title{ANOVA for Randomized Complete Block Design}
\description{
Performs F-statistic based permutation test on data for a randomized complete block design.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:RCBD]{LearnNonparam::RCBD}} -> \code{RCBDANOVA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RCBDANOVA-new}{\code{RCBDANOVA$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="test"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-test'><code>LearnNonparam::PermuTest$test()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RCBDANOVA-new"></a>}}
\if{latex}{\out{\hypertarget{method-RCBDANOVA-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{RCBDANOVA} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RCBDANOVA$new(type = c("permu", "asymp"), n_permu = 0L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate the p-value.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If set to zero (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{RCBDANOVA} object.
}
}
}
