% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCBDTest.R
\name{RCBDTest}
\alias{RCBDTest}
\title{RCBDTest Class}
\description{
Abstract class for tests on samples collected in randomized complete block designs.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{RCBDTest}
}
