% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table1.1.1}
\alias{Table1.1.1}
\title{Sodium Contents}
\format{
An object of class \code{numeric} of length 40.
}
\source{
Table 1.1.1
}
\usage{
Table1.1.1
}
\description{
Sodium contents (in mg) of 40 servings of a food product.
}
\keyword{datasets}
