% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table1.2.1}
\alias{Table1.2.1}
\title{Cycles Until Failure}
\format{
An object of class \code{numeric} of length 20.
}
\source{
Table 1.2.1
}
\usage{
Table1.2.1
}
\description{
The number of cycles (in thousands) that it takes for 20 door latches to fail.
}
\keyword{datasets}
