% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table2.1.1}
\alias{Table2.1.1}
\title{Test Scores}
\format{
An object of class \code{list} of length 2.
}
\source{
Table 2.1.1
}
\usage{
Table2.1.1
}
\description{
Test scores of 7 employees for comparison of methods of instruction.
}
\keyword{datasets}
