% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table2.3.1}
\alias{Table2.3.1}
\title{Runoff Minutes}
\format{
An object of class \code{data.frame} with 8 rows and 2 columns.
}
\source{
Table 2.3.1
}
\usage{
Table2.3.1
}
\description{
The numbers of minutes it took to obtain various amounts of runoff on each plot.
}
\keyword{datasets}
