% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table2.6.1}
\alias{Table2.6.1}
\title{Hours Until Recharge}
\format{
An object of class \code{data.frame} with 4 rows and 2 columns.
}
\source{
Table 2.6.1
}
\usage{
Table2.6.1
}
\description{
The numbers of hours that 2 brands of laptop computers function before battery recharging is necessary.
}
\keyword{datasets}
