% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table2.6.2}
\alias{Table2.6.2}
\title{Cerium Amounts}
\format{
An object of class \code{data.frame} with 6 rows and 2 columns.
}
\source{
Table 2.6.2
}
\usage{
Table2.6.2
}
\description{
The amounts of cerium measured in samples of granite and basalt.
}
\keyword{datasets}
