% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table2.8.1}
\alias{Table2.8.1}
\title{Ounces Of Beverage}
\format{
An object of class \code{data.frame} with 5 rows and 2 columns.
}
\source{
Table 2.8.1
}
\usage{
Table2.8.1
}
\description{
The amounts of liquid in randomly selected beverage containers before and after the filling process has been repaired.
}
\keyword{datasets}
