% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table3.1.2}
\alias{Table3.1.2}
\title{Normal Samples}
\format{
An object of class \code{data.frame} with 5 rows and 3 columns.
}
\source{
Table 3.1.2
}
\usage{
Table3.1.2
}
\description{
Observations randomly sampled from normal populations with means 15, 25 and 30, respectively, and standard deviation 9.
}
\keyword{datasets}
