% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table3.2.2}
\alias{Table3.2.2}
\title{Logarithms of Bacteria Counts}
\format{
An object of class \code{list} of length 4.
}
\source{
Table 3.2.2
}
\usage{
Table3.2.2
}
\description{
Logarithms of counts of bacteria in 4 samples, which respectively were treated with 3 kills and left untreated for the control.
}
\keyword{datasets}
