% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table3.2.3}
\alias{Table3.2.3}
\title{Saltiness Scores}
\format{
An object of class \code{list} of length 3.
}
\source{
Table 3.2.3
}
\usage{
Table3.2.3
}
\description{
Saltiness scores, on a scale of 1 to 5, assigned by a taste expert to samples of 3 food products that differ in the amounts of soymeal they contain.
}
\keyword{datasets}
