% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table3.3.1}
\alias{Table3.3.1}
\title{Percentages of Clay}
\format{
An object of class \code{data.frame} with 6 rows and 4 columns.
}
\source{
Table 3.3.1
}
\usage{
Table3.3.1
}
\description{
The percentages of clay in 6 samples of soil selected from 4 locations.
}
\keyword{datasets}
