% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table3.4.1}
\alias{Table3.4.1}
\title{Phosphorus Contents}
\format{
An object of class \code{data.frame} with 6 rows and 4 columns.
}
\source{
Table 3.4.1
}
\usage{
Table3.4.1
}
\description{
Phosphorus contents of plants under 4 mowing treatments.
}
\keyword{datasets}
