% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table4.1.1}
\alias{Table4.1.1}
\title{Caloric Intake}
\format{
An object of class \code{data.frame} with 5 rows and 2 columns.
}
\source{
Table 4.1.1
}
\usage{
Table4.1.1
}
\description{
The estimated daily caloric intake from dietary information provided using 2 methods by a group of college women.
}
\keyword{datasets}
