% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table4.1.3}
\alias{Table4.1.3}
\title{Cholesterol Reduction}
\format{
An object of class \code{data.frame} with 17 rows and 2 columns.
}
\source{
Table 4.1.3
}
\usage{
Table4.1.3
}
\description{
Reduction in cholesterol after twins were given 2 drugs separately.
}
\keyword{datasets}
