% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table4.4.3}
\alias{Table4.4.3}
\title{Yield Data}
\format{
An object of class \code{data.frame} with 4 rows and 6 columns.
}
\source{
Table 4.4.3
}
\usage{
Table4.4.3
}
\description{
Yield data for a randomized complete block design in which 4 different types of tractors were used in tilling the soil. The blocking factor is location of the fields.
}
\keyword{datasets}
