% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table4.5.3}
\alias{Table4.5.3}
\title{Randomized Complete Block with Ties}
\format{
An object of class \code{data.frame} with 4 rows and 3 columns.
}
\source{
Table 4.5.3
}
\usage{
Table4.5.3
}
\description{
A randomized complete block design with 4 treatments and 3 blocks.
}
\keyword{datasets}
