% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table5.1.2}
\alias{Table5.1.2}
\title{Heterophils and Lymphocytes}
\format{
An object of class \code{data.frame} with 18 rows and 2 columns.
}
\source{
Table5.1.2
}
\usage{
Table5.1.2
}
\description{
Counts of the heterophils and lymphocytes in blood samples from 18 healthy rabbits.
}
\keyword{datasets}
