% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table5.2.2}
\alias{Table5.2.2}
\title{Scores of Projects}
\format{
An object of class \code{data.frame} with 10 rows and 2 columns.
}
\source{
Table5.2.2
}
\usage{
Table5.2.2
}
\description{
Scores of 10 projects at a science fair.
}
\keyword{datasets}
