% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Table5.4.2}
\alias{Table5.4.2}
\title{Satisfaction with Pain-Relief Treatment}
\format{
An object of class \code{data.frame} with 2 rows and 3 columns.
}
\source{
Table5.4.2
}
\usage{
Table5.4.2
}
\description{
Patients' responses with 2 methods of relieving postoperative pain.
}
\keyword{datasets}
