% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleDistributionTest.R
\name{TwoSampleDistributionTest}
\alias{TwoSampleDistributionTest}
\alias{class.distribution}
\title{TwoSampleDistributionTest Class}
\description{
Abstract class for two-sample distribution tests.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{TwoSampleDistributionTest}
}
