% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleLocationTest.R
\name{TwoSampleLocationTest}
\alias{TwoSampleLocationTest}
\alias{class.twosample.location}
\title{TwoSampleLocationTest Class}
\description{
Abstract class for two-sample location tests.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{TwoSampleLocationTest}
}
