% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleTest.R
\name{TwoSampleTest}
\alias{TwoSampleTest}
\title{TwoSampleTest Class}
\description{
This class specializes \code{PermuTest} for two sample permutation tests. Note that it is not recommended to create objects of this class directly.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{TwoSampleTest}
}
