% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANOVA.R
\name{ANOVA}
\alias{ANOVA}
\title{K Sample Permutation Test Based on F Statistic}
\description{
Performs F statistic based k sample permutation test on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::K Sample Permutation Test} -> \code{ANOVA}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ANOVA-new}{\code{ANOVA$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="ggplot"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-ggplot'><code>LearnNonparam::Permutation Test$ggplot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ANOVA-new"></a>}}
\if{latex}{\out{\hypertarget{method-ANOVA-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ANOVA} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ANOVA$new(type = c("permu", "approx"), n_permu = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"permu"} (default) or \code{"approx"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ANOVA} object.
}
}
}
