% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnsariBradley.R
\name{AnsariBradley}
\alias{AnsariBradley}
\alias{twosample.ansari}
\title{Ansari-Bradley Test}
\description{
Performs Ansari-Bradley test on samples.
}
\examples{
pmt(
    "twosample.ansari",
    alternative = "greater", n_permu = 0
)$test(Table2.8.1)$print()

}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{AnsariBradley}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnsariBradley-new}{\code{AnsariBradley$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="test"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-test'><code>LearnNonparam::PermuTest$test()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnsariBradley-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnsariBradley-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{AnsariBradley} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnsariBradley$new(
  type = c("permu", "asymp"),
  alternative = c("two_sided", "less", "greater"),
  n_permu = 10000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate the p-value.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis.}

\item{\code{n_permu}}{an integer indicating number of permutations for the permutation distribution. If set to \code{0}, all permutations will be used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{AnsariBradley} object.
}
}
}
