% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnsariBradley.R
\name{AnsariBradley}
\alias{AnsariBradley}
\title{Ansari-Bradley Test}
\description{
Performs two sample Ansari-Bradley test on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::Two Sample Permutation Test} -> \code{Ansari-Bradley Test}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Ansari-Bradley Test-new}{\code{AnsariBradley$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot'><code>LearnNonparam::Permutation Test$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="print"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-print'><code>LearnNonparam::Permutation Test$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Ansari-Bradley Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Ansari-Bradley Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{AnsariBradley} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnsariBradley$new(
  type = c("exact", "approx", "permu"),
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL,
  conf_level = 0.95
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"exact"} (default), \code{"approx"} or \code{"permu"}. Note that this parameter will be set to \code{"approx"} automatically when there exists ties.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}

\item{\code{conf_level}}{a number specifying confidence level of the interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{AnsariBradley} object.
}
}
}
