% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnsariBradley.R
\name{AnsariBradley}
\alias{AnsariBradley}
\alias{twosample.ansari}
\title{Ansari-Bradley Test}
\description{
Performs two sample Ansari-Bradley test on data vectors.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{AnsariBradley}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AnsariBradley-new}{\code{AnsariBradley$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="test"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-test'><code>LearnNonparam::PermuTest$test()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AnsariBradley-new"></a>}}
\if{latex}{\out{\hypertarget{method-AnsariBradley-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{AnsariBradley} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AnsariBradley$new(
  type = c("permu", "approx"),
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL,
  conf_level = 0.95
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate the p-value.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}

\item{\code{conf_level}}{a number specifying confidence level of the interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{AnsariBradley} object.
}
}
}
