% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ContingencyTableTest.R
\name{ContingencyTableTest}
\alias{ContingencyTableTest}
\title{ContingencyTableTest Class}
\description{
This class specializes \code{PermuTest} for permutation tests for contingency tables. Note that it is not recommended to create objects of this class directly.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{ContingencyTableTest}
}
