% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KolmogorovSmirnov.R
\name{KolmogorovSmirnov}
\alias{KolmogorovSmirnov}
\title{Two Sample Kolmogorov-Smirnov Test}
\description{
Performs two sample Kolmogorov-Smirnov test on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::Two Sample Permutation Test} -> \code{Two Sample Kolmogorov-Smirnov Test}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Two Sample Kolmogorov-Smirnov Test-new}{\code{KolmogorovSmirnov$new()}}
\item \href{#method-Two Sample Kolmogorov-Smirnov Test-plot_ecdfs}{\code{KolmogorovSmirnov$plot_ecdfs()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot_hist"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot_hist'><code>LearnNonparam::Permutation Test$plot_hist()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Two Sample Kolmogorov-Smirnov Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Two Sample Kolmogorov-Smirnov Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{KolmogorovSmirnov} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KolmogorovSmirnov$new(n_permu = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{KolmogorovSmirnov} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Two Sample Kolmogorov-Smirnov Test-plot_ecdfs"></a>}}
\if{latex}{\out{\hypertarget{method-Two Sample Kolmogorov-Smirnov Test-plot_ecdfs}{}}}
\subsection{Method \code{plot_ecdfs()}}{
Draw the empirical cumulative distribution function of each of the data fed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KolmogorovSmirnov$plot_ecdfs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The object itself (invisibly).
}
}
}
