% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KolmogorovSmirnov.R
\name{KolmogorovSmirnov}
\alias{KolmogorovSmirnov}
\alias{twosample.ks}
\title{Two-Sample Kolmogorov-Smirnov Test}
\description{
Performs two-sample Kolmogorov-Smirnov test on samples.
}
\examples{
pmt(
    "twosample.ks", n_permu = 0
)$test(Table2.8.1)$print()

}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{KolmogorovSmirnov}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KolmogorovSmirnov-new}{\code{KolmogorovSmirnov$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="test"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-test'><code>LearnNonparam::PermuTest$test()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KolmogorovSmirnov-new"></a>}}
\if{latex}{\out{\hypertarget{method-KolmogorovSmirnov-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{KolmogorovSmirnov} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KolmogorovSmirnov$new(n_permu = 10000)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n_permu}}{an integer indicating number of permutations for the permutation distribution. If set to zero (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{KolmogorovSmirnov} object.
}
}
}
