% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KSampleTest.R
\name{KSampleTest}
\alias{KSampleTest}
\alias{class.ksample}
\title{KSampleTest Class}
\description{
Abstract class for k-sample tests.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{KSampleTest}
}
