% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCompT.R
\name{MultiCompT}
\alias{MultiCompT}
\title{Multiple Comparison Based on T Test}
\description{
Performs t test based multiple comparison on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::K Sample Permutation Test} -> \code{LearnNonparam::Multiple Comparison} -> \code{Multiple Comparison (t test)}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{bonferroni}}{Whether to apply bonferroni adjustment.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Multiple Comparison (t test)-new}{\code{MultiCompT$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="ggplot"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-ggplot'><code>LearnNonparam::Permutation Test$ggplot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multiple Comparison (t test)-new"></a>}}
\if{latex}{\out{\hypertarget{method-Multiple Comparison (t test)-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MultiCompT} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultiCompT$new(
  type = c("permu", "approx"),
  bonferroni = TRUE,
  signif_level = 0.05,
  n_permu = NULL,
  scoring = c("none", "rank", "vw", "savage")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"permu"} (default) or \code{"approx"}.}

\item{\code{bonferroni}}{a logical indicating whether to apply bonferroni adjustment.}

\item{\code{signif_level}}{a numeric value between zero and one giving the significance level.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}

\item{\code{scoring}}{a character string specifying which scoring system to be used, must be one of \code{"none"} (default), \verb{"rank}, \code{"vw"} or \code{"savage"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{MultiCompT} object.
}
}
}
