% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleComparison.R
\name{MultipleComparison}
\alias{MultipleComparison}
\title{MultipleComparison Class}
\description{
Abstract class for multiple comparisons.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:KSampleTest]{LearnNonparam::KSampleTest}} -> \code{MultipleComparison}
}
