% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleComparison.R
\name{MultipleComparison}
\alias{MultipleComparison}
\title{Multiple Comparison Class}
\description{
This class specializes \code{KSampleTest} for multiple comparisons.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::K Sample Permutation Test} -> \code{Multiple Comparison}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{signif_level}}{The significance level.}

\item{\code{multicomp}}{The multiple comparison result.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Multiple Comparison-new}{\code{MultipleComparison$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot_hist"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot_hist'><code>LearnNonparam::Permutation Test$plot_hist()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Multiple Comparison-new"></a>}}
\if{latex}{\out{\hypertarget{method-Multiple Comparison-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{MultipleComparison} object. Note that it is not recommended to create objects of this class directly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MultipleComparison$new(signif_level = 0.05, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signif_level}}{a numeric value between zero and one giving the significance level.}

\item{\code{...}}{extra parameters passed to \code{PermuTest$initialize}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{MultipleComparison} object.
}
}
}
