% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneSampleTest.R
\name{OneSampleTest}
\alias{OneSampleTest}
\title{OneSampleTest Class}
\description{
This class specializes \code{PermuTest} for one sample tests. Note that it is not recommended to create objects of this class directly.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{OneSampleTest}
}
