% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PermuTest.R
\name{PermuTest}
\alias{PermuTest}
\title{PermuTest Class}
\description{
This is the abstract base class for permutation test objects. Note that it is not recommended to create objects of this class directly.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{type}}{The type of the test.}

\item{\code{method}}{The method used.}

\item{\code{scoring}}{The scoring system used.}

\item{\code{null_value}}{The value of the parameter in the null hypothesis.}

\item{\code{alternative}}{The alternative hypothesis.}

\item{\code{conf_level}}{The confidence level of the interval.}

\item{\code{n_permu}}{The number of permutations used.}

\item{\code{data}}{Data fed into the object.}

\item{\code{data_permu}}{All permutations used.}

\item{\code{statistic}}{The test statistic.}

\item{\code{statistic_permu}}{Test statistics calculated on permutations.}

\item{\code{p_value}}{The p-value.}

\item{\code{estimate}}{The estimated parameter.}

\item{\code{ci}}{The confidence interval.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PermuTest-new}{\code{PermuTest$new()}}
\item \href{#method-PermuTest-feed}{\code{PermuTest$feed()}}
\item \href{#method-PermuTest-print}{\code{PermuTest$print()}}
\item \href{#method-PermuTest-plot}{\code{PermuTest$plot()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermuTest-new"></a>}}
\if{latex}{\out{\hypertarget{method-PermuTest-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{PermuTest} object. Note that it is not recommended to create objects of this class directly.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermuTest$new(
  null_value = 0,
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL,
  conf_level = 0.95,
  scoring = c("none", "rank", "vw", "expon")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{null_value}}{a number specifying the value of the parameter in the null hypothesis.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}

\item{\code{conf_level}}{a number specifying confidence level of the interval.}

\item{\code{scoring}}{a character string specifying which scoring system to be used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{PermuTest} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermuTest-feed"></a>}}
\if{latex}{\out{\hypertarget{method-PermuTest-feed}{}}}
\subsection{Method \code{feed()}}{
Feed the data to the test.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermuTest$feed(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{the data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object itself (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermuTest-print"></a>}}
\if{latex}{\out{\hypertarget{method-PermuTest-print}{}}}
\subsection{Method \code{print()}}{
Print the results of the test. Note that it works only if data has been fed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermuTest$print(digits = getOption("digits"))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{digits}}{number of significant digits to be used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object itself (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PermuTest-plot"></a>}}
\if{latex}{\out{\hypertarget{method-PermuTest-plot}{}}}
\subsection{Method \code{plot()}}{
Plot histogram(s) of the permutation distribution. Note that it works only if the test's type is \code{"permu"}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PermuTest$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{extra parameters passed to \code{ggplot2::stat_bin}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The object itself (invisibly).
}
}
}
