% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt.R
\name{pmt}
\alias{pmt}
\title{Create a test object conveniently}
\usage{
pmt(key, ...)
}
\arguments{
\item{key}{key corresponding to the desired test class. Check \code{pmts} to see all keys available.}

\item{...}{extra parameters passed to the initialize method of the test class.}
}
\description{
This class specializes \code{PermuTest} for permutation tests for contingency tables.
}
