% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt.R
\docType{data}
\name{pmt}
\alias{pmt}
\alias{tests}
\alias{pmts}
\title{Syntactic Sugar for Object Construction}
\format{
An object of class \code{list} of length 22.
}
\usage{
tests

pmt(key, ...)

pmts(
  which = c("all", "onesample", "twosample", "ksample", "multicomp", "paired", "rcbd",
    "association", "table")
)
}
\arguments{
\item{key}{a character string corresponding to the desired test. Check \code{pmts()} to see available keys.}

\item{...}{extra parameters passed to the \code{new()} method of the test class.}

\item{which}{a character string specifying which tests to show. If \code{"all"} (default) then available tests are shown.}
}
\description{
Create a test object conveniently.
}
\keyword{datasets}
