% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt.R
\name{pmt}
\alias{pmt}
\alias{pmts}
\title{Syntactic Sugar for Object Construction}
\usage{
pmt(key, ...)

pmts(
  which = c("onesample", "twosample", "ksample", "multicomp", "paired", "rcbd",
    "association", "table", "all")
)
}
\arguments{
\item{key}{a character string corresponding to the desired test. Check \code{pmts()} for valid keys.}

\item{...}{extra parameters passed to the constructor.}

\item{which}{a character string specifying the desired tests.}
}
\description{
Construct test objects in a unified way.
}
