% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantile.R
\name{Quantile}
\alias{Quantile}
\title{Quantile Test}
\description{
Performs one sample quantile test on data vectors. In addition, a confidence interval will be calculated.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::One Sample Test} -> \code{Quantile Test}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{p}}{The probability.}

\item{\code{null_value}}{The true value of the quantile.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Quantile Test-new}{\code{Quantile$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="ggplot"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-ggplot'><code>LearnNonparam::Permutation Test$ggplot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quantile Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Quantile Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Quantile} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quantile$new(
  p = 0.5,
  null_value = 0,
  alternative = c("two_sided", "less", "greater"),
  conf_level = 0.95
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{p}}{a numeric between 0 and 1 indicating the probability.}

\item{\code{null_value}}{a number indicating the true value of the quantile.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}

\item{\code{conf_level}}{a number specifying confidence level of the interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Quantile} object.
}
}
}
