% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Quantile.R
\name{Quantile}
\alias{Quantile}
\alias{onesample.quantile}
\title{Quantile Test}
\description{
Performs one sample quantile test on data vectors. In addition, a confidence interval will be calculated.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:OneSampleTest]{LearnNonparam::OneSampleTest}} -> \code{Quantile}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{prob}}{The probability.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Quantile-new}{\code{Quantile$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="feed"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-feed'><code>LearnNonparam::PermuTest$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Quantile-new"></a>}}
\if{latex}{\out{\hypertarget{method-Quantile-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Quantile} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Quantile$new(
  prob = 0.5,
  null_value = 0,
  alternative = c("two_sided", "less", "greater"),
  conf_level = 0.95
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prob}}{a numeric between 0 and 1 indicating the probability.}

\item{\code{null_value}}{a number specifying the value of the parameter in the null hypothesis.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis.}

\item{\code{conf_level}}{a number specifying confidence level of the interval.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Quantile} object.
}
}
}
