% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCBD.R
\name{RCBD}
\alias{RCBD}
\title{RCBD Class}
\description{
This class specializes \code{PermuTest} for randomized complete block design. Note that it is not recommended to create objects of this class directly.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{RCBD}
}
