% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCBDANOVA.R
\name{RCBDANOVA}
\alias{RCBDANOVA}
\title{ANOVA for Randomized Complete Block Design}
\description{
Performs F-statistic based permutation test on data for a randomized complete block design.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::Randomized Complete Block Design} -> \code{ANOVA for Randomized Complete Block Design}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ANOVA for Randomized Complete Block Design-new}{\code{RCBDANOVA$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot_hist"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot_hist'><code>LearnNonparam::Permutation Test$plot_hist()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ANOVA for Randomized Complete Block Design-new"></a>}}
\if{latex}{\out{\hypertarget{method-ANOVA for Randomized Complete Block Design-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{RCBDANOVA} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RCBDANOVA$new(type = c("permu", "approx"), n_permu = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"permu"} (default) or \code{"approx"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{RCBDANOVA} object.
}
}
}
