% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScoreSum.R
\name{ScoreSum}
\alias{ScoreSum}
\title{Score Sum Test}
\description{
Performs two sample score sum test on data vectors, which is almost the same as wilcoxon rank sum test but uses not only rank but also scoring systems other than it.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{ScoreSum}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScoreSum-new}{\code{ScoreSum$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="feed"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-feed'><code>LearnNonparam::PermuTest$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreSum-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreSum-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{ScoreSum} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreSum$new(
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL,
  scoring = c("rank", "vw", "expon")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}

\item{\code{scoring}}{a character string specifying which scoring system to be used, must be one of \code{"rank"} (default), \code{"vw"} or \code{"expon"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{ScoreSum} object.
}
}
}
