% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiegelTukey.R
\name{SiegelTukey}
\alias{SiegelTukey}
\title{Siegel-Tukey Test}
\description{
Performs two sample Siegel-Tukey test on data vectors.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{\link[LearnNonparam:Wilcoxon]{LearnNonparam::Wilcoxon}} -> \code{SiegelTukey}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SiegelTukey-new}{\code{SiegelTukey$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="feed"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-feed'><code>LearnNonparam::PermuTest$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SiegelTukey-new"></a>}}
\if{latex}{\out{\hypertarget{method-SiegelTukey-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SiegelTukey} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiegelTukey$new(adjust_median = FALSE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adjust_median}}{a logical indicating whether the median difference between groups is levelled before the test is conducted.}

\item{\code{...}}{extra parameters passed to \code{Wilcoxon$new()}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{SiegelTukey} object.
}
}
}
