% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SiegelTukey.R
\name{SiegelTukey}
\alias{SiegelTukey}
\title{Siegel-Tukey Test}
\description{
Performs two sample Siegel-Tukey test on data vectors.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::Two Sample Permutation Test} -> \code{LearnNonparam::Two Sample Wilcoxon Test} -> \code{Siegel-Tukey Test}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Siegel-Tukey Test-new}{\code{SiegelTukey$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="ggplot"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-ggplot'><code>LearnNonparam::Permutation Test$ggplot()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Siegel-Tukey Test-new"></a>}}
\if{latex}{\out{\hypertarget{method-Siegel-Tukey Test-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SiegelTukey} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SiegelTukey$new(
  adjust_median = FALSE,
  ...,
  alternative = c("two_sided", "less", "greater")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{adjust_median}}{a logical indicating whether the median difference between groups is levelled before the test is conducted.}

\item{\code{...}}{extra parameters passed to \code{Wilcoxon$new()}.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{SiegelTukey} object.
}
}
}
