% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sign.R
\name{Sign}
\alias{Sign}
\title{Sign Test}
\description{
Performs two sample sign test on data vectors.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{\link[LearnNonparam:TwoSamplePairedTest]{LearnNonparam::TwoSamplePairedTest}} -> \code{Sign}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Sign-new}{\code{Sign$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="feed"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-feed'><code>LearnNonparam::PermuTest$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Sign-new"></a>}}
\if{latex}{\out{\hypertarget{method-Sign-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{Sign} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Sign$new(
  type = c("exact", "approx", "permu"),
  correct = TRUE,
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"exact"} (default), \code{"approx"} or \code{"permu"}.}

\item{\code{correct}}{a logical indicating whether to apply continuity correction in the normal approximation for the p-value.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{Sign} object.
}
}
}
