% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignedScore.R
\name{SignedScore}
\alias{SignedScore}
\alias{paired.signedscore}
\title{Signed Score Test}
\description{
Performs two sample signed score test on data vectors.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{\link[LearnNonparam:TwoSamplePairedTest]{LearnNonparam::TwoSamplePairedTest}} -> \code{SignedScore}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SignedScore-new}{\code{SignedScore$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="feed"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-feed'><code>LearnNonparam::PermuTest$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SignedScore-new"></a>}}
\if{latex}{\out{\hypertarget{method-SignedScore-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{SignedScore} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SignedScore$new(
  type = c("permu", "approx"),
  correct = TRUE,
  ranking_method = c("with_zeros", "ignore"),
  alternative = c("two_sided", "less", "greater"),
  n_permu = NULL,
  scoring = c("rank", "vw", "expon")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"permu"} (default) or \code{"approx"}.}

\item{\code{correct}}{a logical indicating whether to apply continuity correction in the normal approximation for the p-value when \code{scoring} is set to \code{"rank"}.}

\item{\code{ranking_method}}{a character string specifying the method of ranking data in computing adjusted signed ranks for tied data, must be one of \code{"with_zeros"} (default) or \code{"ignore"}. Note that the data fed will be modified when this parameter is set to \code{"ignore"}.}

\item{\code{alternative}}{a character string specifying the alternative hypothesis, must be one of \code{"two_sided"} (default), \code{"greater"} or \code{"less"}.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}

\item{\code{scoring}}{a character string specifying which scoring system to be used, must be one of \code{"rank"} (default), \code{"vw"} or \code{"expon"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{SignedScore} object.
}
}
}
