% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyHSD.R
\name{TukeyHSD}
\alias{TukeyHSD}
\alias{multicomp.tukey}
\title{Tukey's HSD}
\description{
Performs Tukey's honest significant difference method on data vectors. Procedure modified by Kramer is performed for unequal sample sizes.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:KSampleTest]{LearnNonparam::KSampleTest}} -> \code{\link[LearnNonparam:MultipleComparison]{LearnNonparam::MultipleComparison}} -> \code{TukeyHSD}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TukeyHSD-new}{\code{TukeyHSD$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="plot"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-plot'><code>LearnNonparam::PermuTest$plot()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="print"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-print'><code>LearnNonparam::PermuTest$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="PermuTest" data-id="test"><a href='../../LearnNonparam/html/PermuTest.html#method-PermuTest-test'><code>LearnNonparam::PermuTest$test()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TukeyHSD-new"></a>}}
\if{latex}{\out{\hypertarget{method-TukeyHSD-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TukeyHSD} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TukeyHSD$new(
  type = c("permu", "asymp"),
  scoring = c("none", "rank", "vw", "expon"),
  conf_level = 0.95,
  n_permu = 0L
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate the p-value.}

\item{\code{scoring}}{a character string specifying which scoring system to be used.}

\item{\code{conf_level}}{a numeric value between zero and one giving the family-wise confidence level to use.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If set to zero (default) then all permutations are used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{TukeyHSD} object.
}
}
}
