% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TukeyHSD.R
\name{TukeyHSD}
\alias{TukeyHSD}
\title{Tukey's HSD}
\description{
Performs Tukey's honest significant difference method on data vectors. Note that procedure modified by Kramer is performed for unequal sample sizes.
}
\section{Super classes}{
\code{LearnNonparam::Permutation Test} -> \code{LearnNonparam::K Sample Permutation Test} -> \code{LearnNonparam::Multiple Comparison} -> \code{Tukey HSD}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Tukey HSD-new}{\code{TukeyHSD$new()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="feed"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-feed'><code>LearnNonparam::Permutation Test$feed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="LearnNonparam" data-topic="Permutation Test" data-id="plot_hist"><a href='../../LearnNonparam/html/Permutation Test.html#method-Permutation Test-plot_hist'><code>LearnNonparam::Permutation Test$plot_hist()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Tukey HSD-new"></a>}}
\if{latex}{\out{\hypertarget{method-Tukey HSD-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TukeyHSD} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TukeyHSD$new(
  type = c("permu", "approx"),
  conf_level = 0.95,
  n_permu = NULL,
  scoring = c("none", "rank", "vw", "expon")
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{a character string specifying the way to calculate p-values, must be one of \code{"permu"} (default) or \code{"approx"}.}

\item{\code{conf_level}}{a numeric value between zero and one giving the family-wise confidence level to use.}

\item{\code{n_permu}}{an integer specifying how many permutations should be used to construct the permutation distribution. If \code{NULL} (default) then all permutations are used.}

\item{\code{scoring}}{a character string specifying which scoring system to be used, must be one of \code{"none"} (default), \verb{"rank}, \code{"vw"} or \code{"expon"}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{TukeyHSD} object.
}
}
}
