% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleAssociationTest.R
\name{TwoSampleAssociationTest}
\alias{TwoSampleAssociationTest}
\alias{class.association}
\title{TwoSampleAssociationTest Class}
\description{
Abstract class for two-sample association tests.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{\link[LearnNonparam:TwoSamplePairedTest]{LearnNonparam::TwoSamplePairedTest}} -> \code{TwoSampleAssociationTest}
}
