% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSamplePairedTest.R
\name{TwoSamplePairedTest}
\alias{TwoSamplePairedTest}
\alias{class.paired}
\title{TwoSamplePairedTest Class}
\description{
Abstract class for paired two-sample tests.
}
\section{Super classes}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{\link[LearnNonparam:TwoSampleTest]{LearnNonparam::TwoSampleTest}} -> \code{TwoSamplePairedTest}
}
