% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoSampleTest.R
\name{TwoSampleTest}
\alias{TwoSampleTest}
\alias{class.twosample}
\title{TwoSampleTest Class}
\description{
Abstract class for two-sample tests.
}
\section{Super class}{
\code{\link[LearnNonparam:PermuTest]{LearnNonparam::PermuTest}} -> \code{TwoSampleTest}
}
